backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-redis==6.0.0
django-environ==0.12.0
django-ckeditor-5==0.2.18

pillow==11.3.0

uvicorn==0.34.3

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.9.0

psycopg2==2.9.10

